Update `check-cla` to customize which CLA repo to use (#91)
Extract `set-commit-status` into a new action.

Reworks `check-cla` to automatically open the CLA PR when a missing contributor is detected. This
simplifies the CLA verification process so maintainers only need to manually match the signed PDF to
the PR and click merge. This will theoretically also be more visible to the user.